+++ src/buttondevteam/core/TBMCPlayer.java
@@ -1,4 +1,4 @@
-package buttondevteam.core.player;
+package buttondevteam.core;
 
 import java.io.File;
 import java.io.IOException;
