+++ src/buttondevteam/core/TBMCPlayerSaveEvent.java
@@ -1,4 +1,4 @@
-package buttondevteam.core.player;
+package buttondevteam.core;
 
 import org.bukkit.configuration.file.YamlConfiguration;
 import org.bukkit.event.Event;
